Kits in career events
Bugfix: if you win a career event and get a kit as a prize (for example, stroker kit for a V10), then you won't be able to install it because parts in its inventory are listed in a reverse order and not properly copied into player inventory. This commit contains a small fix for to avoid that.